backend/requirements.txt
# MediExplain AI - Requirements
# Install with: pip install -r requirements.txt

# ── Web Framework ─────────────────────────────────────────────
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
starlette==0.27.0

# ── Pydantic (Data Validation) ────────────────────────────────
pydantic==2.5.0
pydantic-core==2.14.1
annotated-types==0.6.0

# ── HTTP & Requests ───────────────────────────────────────────
requests==2.31.0
httpx==0.25.1
certifi==2023.11.17
charset-normalizer==3.3.2
idna==3.6
urllib3==2.1.0

# ── OCR & Image Processing ────────────────────────────────────
pytesseract==0.3.10
Pillow==10.1.0
pdf2image==1.16.3
PyPDF2==3.0.1
opencv-python==4.8.1.78

# ── Translation ───────────────────────────────────────────────
deep-translator==1.11.4

# ── Database (Optional - for saving reports) ─────────────────
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# ── Redis & Caching (Optional) ───────────────────────────────
redis==5.0.1

# ── File Processing ───────────────────────────────────────────
openpyxl==3.1.2
python-docx==1.1.0

# ── Utilities ─────────────────────────────────────────────────
python-dotenv==1.0.0
click==8.1.7
colorama==0.4.6
typing-extensions==4.8.0

# ── Async ─────────────────────────────────────────────────────
anyio==3.7.1
h11==0.14.0

# ── Testing ───────────────────────────────────────────────────
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.1

# ──────────────────────────────────────────────────────────────
# OPTIONAL - Advanced NLP (install separately if needed)
# ──────────────────────────────────────────────────────────────
# spacy==3.6.1
# scispacy==0.5.3
# torch==2.1.1
# transformers==4.35.2
#
# After installing spacy, run:
# pip install https://s3-us-west-2.amazonaws.com/ai2-s3-scispacy/releases/v0.5.1/en_core_sci_sm-0.5.1.tar.gz
# ──────────────────────────────────────────────────────────────
